+++ commands/fun/action.js
@@ -227,7 +227,7 @@ module.exports = {
   
       // Fetch function
       async function apiFetch(action) {
-        await neko.sfw[action]()
+        await neko[action]()
           .then(data => url = data.url)
       }
     },