+++ commands/fun/animal.js
@@ -18,7 +18,7 @@ module.exports = {
 
     // Fetch function
     async function apiFetch(action) {
-      await neko.sfw[action]()
+      await neko[action]()
         .then(data => url = data.url)
     }