+++ commands/fun/neko.js
@@ -12,7 +12,7 @@ module.exports = {
     const neko = new nekos();
 
     let url;
-    await neko.sfw.neko()
+    await neko.neko()
       .then(data => url = data.url);
 
     const nekoEmbed = {