+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Removed un-needed functions from utils.js
 * Renamed utilities.js to utils.js
 * Updated sb!info so it shows uptime correctly
+* Updated commands that ran off nekos.life database to v3 (they removed the nsfw ones :/)
 
 ## 2.2.4 - 09/04/2022
 * StenBot no longer crashes when joining a new server (yes, it took this long to fix the bug)